refactor(gh-364): shift `inference/flowMChandler.py` and `inference/utils.py` to kokab (#468)
* refactor(gh-364): shift `inference/flowMChandler.py` and `inference/utils.py` to kokab

* refactor: update sage scripts; catering the changes made in 429c60a0dc2e3bddcd1afd68282dc43b70af4a42
