+++ src/kokab/chi_eff_q_relation/sage.py
@@ -21,7 +21,7 @@ from glob import glob
 from jax import numpy as jnp, random as jrd
 from numpyro.distributions import Uniform
 
-from gwkokab.inference import Bake, flowMChandler, PoissonLikelihood
+from gwkokab.inference import Bake, PoissonLikelihood
 from gwkokab.logger import enable_logging
 from gwkokab.models import ChiEffMassRatioCorrelated
 from gwkokab.parameters import (
@@ -41,6 +41,7 @@ from kokab.utils.common import (
     read_json,
     vt_json_read_and_process,
 )
+from kokab.utils.flowMC_helper import flowMChandler
 
 
 class RedshiftReferencePrior(Uniform):
