+++ src/kokab/ecc_matters/sage.py
@@ -21,7 +21,7 @@ from glob import glob
 import numpy as np
 from jax import random as jrd
 
-from gwkokab.inference import Bake, flowMChandler, PoissonLikelihood
+from gwkokab.inference import Bake, PoissonLikelihood
 from gwkokab.logger import enable_logging
 from gwkokab.parameters import ECCENTRICITY, PRIMARY_MASS_SOURCE, SECONDARY_MASS_SOURCE
 from gwkokab.poisson_mean import PoissonMean
@@ -35,6 +35,7 @@ from kokab.utils.common import (
     read_json,
     vt_json_read_and_process,
 )
+from kokab.utils.flowMC_helper import flowMChandler
 
 
 def make_parser() -> ArgumentParser:
