+++ src/kokab/n_pls_m_gs/sage.py
@@ -23,7 +23,7 @@ import numpy as np
 from jax import random as jrd
 
 import gwkokab
-from gwkokab.inference import Bake, flowMChandler, PoissonLikelihood
+from gwkokab.inference import Bake, PoissonLikelihood
 from gwkokab.logger import enable_logging
 from gwkokab.models import NPowerlawMGaussian
 from gwkokab.models.utils import create_truncated_normal_distributions
@@ -48,6 +48,7 @@ from kokab.utils.common import (
     read_json,
     vt_json_read_and_process,
 )
+from kokab.utils.flowMC_helper import flowMChandler
 
 
 def make_parser() -> ArgumentParser:
