+++ src/kokab/one_powerlaw_one_peak/sage.py
@@ -21,11 +21,7 @@ from glob import glob
 import numpy as np
 from jax import random as jrd
 
-from gwkokab.inference import (
-    Bake,
-    flowMChandler,
-    PoissonLikelihood,
-)
+from gwkokab.inference import Bake, PoissonLikelihood
 from gwkokab.logger import enable_logging
 from gwkokab.models import SmoothedPowerlawAndPeak
 from gwkokab.parameters import (
@@ -42,6 +38,7 @@ from kokab.utils.common import (
     read_json,
     vt_json_read_and_process,
 )
+from kokab.utils.flowMC_helper import flowMChandler
 
 
 def make_parser() -> ArgumentParser:
